Move create simulation months to ochre simulation
